LayoutTests/http/tests/resourceLoadStatistics/non-sandboxed-nesting-iframe-with-non-sandboxed-iframe-redirect-localhost-to-ip-to-localhost-expected.txt

Tests that redirects for a non-sandboxed iframe nested in a non-sandboxed iframe get counted properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testRunner.isStatisticsRegisteredAsSubFrameUnder("http://localhost", "http://127.0.0.1") is true
PASS testRunner.isStatisticsRegisteredAsRedirectingTo("http://localhost", "http://127.0.0.1") is true
PASS testRunner.isStatisticsRegisteredAsRedirectingTo("http://127.0.0.1", "http://localhost") is true
PASS successfullyParsed is true

TEST COMPLETE
